Add tests to correct file.
